ci: use warnings-as-errors on Windows+MSVC+Bazel
Enable most warnings (/W3) and treat warnings as errors (/WX) on the
Windows+MSVC+Bazel builds. We already do this on the Windows+MSVC+CMake
builds, but those are post-submit and therefore break from time to time.
Note that only the CMake builds compile for x86, so there are some
warnings we miss anyway.

The change is more intrusive than just adding some --copt options
because we only want *our* code compiled with these options, we do not
want to hear about warnings in our dependencies.

+++ google/cloud/storage/internal/grpc_client.cc
@@ -919,7 +919,7 @@ google::storage::v1::Bucket::Cors GrpcClient::ToProto(CorsEntry const& rhs) {
     result.add_response_header(v);
   }
   if (rhs.max_age_seconds.has_value()) {
-    result.set_max_age_seconds(*rhs.max_age_seconds);
+    result.set_max_age_seconds(static_cast<std::int32_t>(*rhs.max_age_seconds));
   }
   return result;
 }
